Refactor the project and the docker compose file
